backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pandas==2.2.3
openpyxl==3.1.5
motor==3.6.0
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
